+++ MeltingApp/MeltingApp/Views/Pages/CreateEvent.xaml.cs
@@ -9,7 +9,7 @@ namespace MeltingApp.Views.Pages
 	{
 		public CreateEvent ()
 		{
-		    BindingContext = new EventViewModel();
+		    //BindingContext = new EventViewModel();
             InitializeComponent();
             NavigationPage.SetHasNavigationBar(this, false);
         }
