+++ MeltingApp/MeltingApp/Views/Pages/MainPage.xaml
@@ -12,7 +12,6 @@
             <Button Text="Help" Command="{Binding NavigateToHelpPageCommand}"></Button>
             <Button Text="Static Info" Command="{Binding NavigateToStaticInfoPage}"></Button>
             <Button Text="Finder" Command="{Binding NavigateToFinderPage}"></Button>
-            <Button Text="Create Event" Command="{Binding NavigateToCreateEventPageCommand}"></Button>
             <Button Text="Get All Events" Command="{Binding NavigateToEventViewModelCommand}"></Button>
             <Button Text="View Event" Command="{Binding NavigateToViewEventPageCommand}"></Button>
             <Button x:Name="ViewButton" Text="View Profile" Command="{Binding NavigateToProfileViewModelCommand}"></Button>
